+++ app/page.tsx
@@ -1,15 +1,18 @@
-import About from "@/components/About/About";
-import Name from "@/components/Name/Name";
-import SectionDivider from "@/components/SectionDivider/SectionDivider";
+import About from "@/components/Sections/About/About";
+import Name from "@/components/Sections/Name/Name";
+import SectionDivider from "@/components/Common/SectionDivider/SectionDivider";
+import WorkExperience from "@/components/Sections/WorkExperience/WorkExperience";
 
 export default function Home() {
   return (
     <>
       <Name />
+      
       <SectionDivider sectionTitle="About Me" />
       <About />
 
       <SectionDivider sectionTitle="Work Experience" />
+      <WorkExperience />
     </>
   )
 }
