+++ components/Common/SkillBadge/SkillBadge.tsx
@@ -1,4 +1,4 @@
-import { Badge } from "../ui/badge";
+import { Badge } from "@/components/ui/badge";
 
 type skillBadgeProps = {
     skill: string;
