+++ components/Sections/WorkExperience/WorkExperience.tsx
@@ -1,0 +1,8 @@
+export default function WorkExperience() {
+    return (
+        <div className="w-max mx-auto my-20 text-center">
+            <div className="text-4xl font-medium">Work Experience</div>
+            <div className="text-lg text-muted-foreground">test</div>
+        </div>
+    )
+}
